Building model with GMMHMM as per paper told
However, the input of fit method is only one, which means it may be the combination of observed state (occupancy) and emission (all_feature_reduced). It gives error as bellow:

ValueError: GMM expects at least 2 samples provided in a 2 dimensional array-like input

it seems like the value of all_feature_reduced. SO the value of y_train:

[ 1.  1.  0.  1.  1.  0.  0.  0.  1.  1.  1.  0.  1.  0.  1.  0.  0.  1.
  1.  0.  1.  1.  0.  0.  0.  1.  0.  0.  1.  1.  0.  0.  0.  1.  0.  1.
  1.  0.  1.  1.  1.  1.  1.  1.  1.  0.  0.  0.  1.  1.  1.  0.  1.  0.
  0.  0.  1.  0.  1.  1.  1.  1.  1.  1.  1.  0.  1.  0.  0.  0.  0.  0.
  1.  0.  0.  0.]

mus be changed into:
[ [1.], [ 1.],  ........ ]